Creating BankSlip ORM
